+++ src/js/const/TODO.js
@@ -1,0 +1,5 @@
+export const FILTER = {
+  ACTIVE: 'ACTIVE',
+  ALL: 'ALL',
+  COMPLETED: 'COMPLETED', 
+}
