+++ src/js/store/users.js
@@ -1,0 +1,5 @@
+const INIT_STATE = {
+  users: [],
+  todos: [],
+  currentFilter: 
+}
