+++ src/js/utils/common.js
@@ -1,0 +1,1 @@
+export const clone = value => JSON.parse(JSON.stringify(value));
